Add prompt for setting params.external_domain
The server cert is now generated with the external_domain as a SAN, but
as it is generated on completion of the new wizard, it gets generated
before the user can add the external domain to their environment.

With this change, the external doomain, if needed, is put in the env
file prior to the certs being generated.